Added datastore.DeleteGame(roomCode) function to allow room brokers to delete a finished game from the redis datatstore when the last active player disconnects.
Added a Cache-Control header to disable caching of game pages to
support the stale game cleanup, otherwise the cached copy would
make a request that would create a new state for the stale game.

The scope of efficiently managing server resources is limited in this
project, but we have considered it and attempted to make some effort
in this regard.